jeyn_core/workflow_controller/requirements.txt
django~=3.2
djangorestframework ~=3.12
dapr~=1.2.0
requests~=2.25
psycopg2-binary==2.9.1